Version 0.9.0 Available

+++ R/fire_rate_distribution.R
@@ -26,9 +26,9 @@
 #' @examples
 #' ## Load relavent data
 #' data("fire_data")
-#' aoi <- read_sf(system.file("extdata","extdata.gpkg",package="BurnP3.HelpR"),"aoi")
+#' aoi <- sf::st_read(system.file("extdata","extdata.gpkg",package="BurnP3.HelpR"),"aoi")
 #' output_location <- paste0(tempdir(),"\\")
-#' zones <- rast(system.file("extdata","zones.tif",package="BurnP3.HelpR"))
+#' zones <- terra::rast(system.file("extdata","zones.tif",package="BurnP3.HelpR"))
 #' data("season_df")
 #' zone_names = c("Alpine-E","Montane-E","Alpine-W","Montane-W","IDF")
 #'
@@ -37,8 +37,8 @@
 #'                              aoi = aoi,
 #'                              output_location = output_location,
 #'                              date_col = "REP_DATE",
-#'                              seasonal = F,
-#'                              zonal = F,
+#'                              seasonal = FALSE,
+#'                              zonal = FALSE,
 #'                              seasons = "",
 #'                              zones,
 #'                              zone_names = "",
@@ -52,8 +52,8 @@
 #'                              aoi = aoi,
 #'                              output_location = output_location,
 #'                              date_col = "REP_DATE",
-#'                              seasonal = T,
-#'                              zonal = F,
+#'                              seasonal = TRUE,
+#'                              zonal = FALSE,
 #'                              seasons = season_df,
 #'                              zones,
 #'                              zone_names = "",
@@ -67,8 +67,8 @@
 #'                              aoi = aoi,
 #'                              output_location = output_location,
 #'                              date_col = "REP_DATE",
-#'                              seasonal = F,
-#'                              zonal = T,
+#'                              seasonal = FALSE,
+#'                              zonal = TRUE,
 #'                              seasons = "",
 #'                              zones = zones,
 #'                              zone_names = zone_names,
@@ -78,7 +78,7 @@
 #'
 #' print(fr)
 #'
-#' unlink(output_location, recursive = T)
+#' unlink(output_location)
 #'
 
 fire_rate_distribution <- function(input, date_col, date_format = "%Y/%m/%d", aoi, output_location, seasonal=F, zonal=F, seasons = "", zones, zone_names = "", min_fire_size = 0.01, causes = c("H","L")){
@@ -110,7 +110,7 @@ fire_rate_distribution <- function(input, date_col, date_format = "%Y/%m/%d", ao
   }
 
   if (zonal) {
-    input$zone <- terra::extract(zones,input)
+    input$zone <- terra::extract(project(zones,crs(input),method="near"),input)$zones
     if (length(which(is.na(input$zone))) > 0) {input <- input[-which(is.na(input$zone)),]}
   }
 
@@ -126,13 +126,14 @@ fire_rate_distribution <- function(input, date_col, date_format = "%Y/%m/%d", ao
   )
 
   colnames(fire_rate) <- tolower(colnames(fire_rate))
-  print(data.frame(cause = fire_rate$cause, numeric_cause = as.numeric(as.factor(as.character(fire_rate$cause)))))
   fire_rate$cause <- as.numeric(as.factor(as.character(fire_rate$cause)))
-  if (output_location == dir_list[[2]]) {
+  print(fire_rate)
+  if (exists("dir_list") && output_location == dir_list[[2]]) {
     write.csv(fire_rate,
               paste0(output_location,"Inputs/2. Modules/Distribution Tables/Fire_Rate_Distribution.csv"),
-              row.names = F)} else {
-                write.csv(fire_rate, paste0(output_location,"Fire_Rate_Distribution.csv"),row.names = F)
-              }
+              row.names = F)
+    } else {
+        write.csv(fire_rate, paste0(output_location,"Fire_Rate_Distribution.csv"),row.names = F)
+        }
   return(fire_rate)
 }
